api/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
eventlet==0.30.2
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
python-socketio==5.3.0
python-engineio==4.1.0
psycopg2-binary==2.9.6
python-dotenv==0.21.0
SQLAlchemy==1.4.49
argon2-cffi==23.1.0
alembic==1.11.1
requests==2.28.2